Moved file creation to db service.
