+++ HCCGo/app/js/dbService.js
@@ -5,22 +5,71 @@ dbService = angular.module('dbService', []);
  they allow the loading of the DB's only once, which stops any race conditions of trying to open the DB more than once.
  *
  * The DBs returned by the `dbService` are all nedb Databases.
- * 
+ *
  * Documentation on [Nedb](https://github.com/louischatriot/nedb).
- * 
+ *
  * @ngdoc service
  * @memberof HCCGo
  * @class dbService
  */
-dbService.service('dbService', ['filePathService', function(filePathService) {
+dbService.service('dbService', ['filePathService', '$log', function(filePathService, $log) {
 
   const DataStore = require('nedb');
+  var path = require('path');
+  var dataPath = filePathService.getDataPath();
+  var jobHistory = path.join(__dirname, 'data/jobHistory.json');
   var jobHistoryPath = filePathService.getJobHistory();
-  var jobHistoryDB = new DataStore({ filename: jobHistoryPath, autoload: true });
+  var jobHistoryDB;
   var submittedJobsPath = filePathService.getSubmittedJobs();
-  var submittedJobsDB = new DataStore({ filename: submittedJobsPath, autoload: true });
+  var submittedJobsDB;
+
+  // Check if app data folder is there, if not, create one with default json file
+  var fs = require('fs');
+  fs.exists(dataPath, function(exists) {
+    if(!exists) {
+      // folder doesn't exist
+      fs.mkdir(dataPath, function() {
+          // create default files
+          fs.createWriteStream(jobHistoryPath);
+          jobHistoryDB = new DataStore({ filename: jobHistoryPath, autoload: true });
+          $.getJSON(jobHistory, function(json) {
+            jobHistoryDB.insert(json.jobs[0], function(err, newDoc) {
+              if(err) $log.err(err);
+            });
+          });
+          fs.createWriteStream(submittedJobsPath);
+          submittedJobsDB = new DataStore({ filename: submittedJobsPath, autoload: true });
+      });
+    }
+    else {
+      // folder does exist
+      fs.exists(jobHistoryPath, function(fileExists) {
+        if(!fileExists) {
+          // jobhistory file doesn't exist
+          fs.createWriteStream(jobHistoryPath);
+          jobHistoryDB = new DataStore({ filename: jobHistoryPath, autoload: true });
+          $.getJSON(jobHistory, function(json) {
+            jobHistoryDB.insert(json.jobs[0], function(err, newDoc) {
+              if(err) $log.err(err);
+            });
+          });
+        }
+        else {
+          jobHistoryDB = new DataStore({ filename: jobHistoryPath, autoload: true });
+        }
+      });
+      fs.exists(submittedJobsPath, function(fileExists) {
+        if(!fileExists) {
+          // jobsubmission file doesn't exist
+          fs.createWriteStream(submittedJobsPath);
+        }
+        submittedJobsDB = new DataStore({ filename: submittedJobsPath, autoload: true });
+      });
+    }
+  });
+
   return {
-    
+
     /**
      * Get the Job History DB reference.
      * @method getJobHistoryDB
@@ -30,7 +79,7 @@ dbService.service('dbService', ['filePathService', function(filePathService) {
     getJobHistoryDB: function() {
       return jobHistoryDB;
     },
-    
+
     /**
      * Get the Submitted jobs DB reference.
      * @method getSubmittedJobsDB
